attendance_system/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-MySQLdb==1.0.1
opencv-python==4.8.0.74
opencv-contrib-python==4.8.0.74
numpy==1.24.3
python-dotenv==1.0.0